Add reverse playback logic for storing and playing quicksounds. Refactor cogs
