+++ cogs/Entrance.py
@@ -8,6 +8,8 @@ from discord.ext import commands
 from modules.database import GetDB
 from modules.app_config import config
 from modules.player import player
+from modules.errors import *
+from modules.helper_functions import *
 from modules.aliases import sounds
 
 # mapping of mm/dd to holidays
@@ -78,6 +80,83 @@ class Entrance(commands.Cog):
         else:
             return
 
+    @commands.group()
+    @commands.guild_only()
+    async def entrance(self, ctx):
+        """
+        Plays a sound when you enter a voice channel to announce your entry. 
+        Will not play unless you have been out of voice for more than an hour since your last entry has played.
+        """
+
+    @entrance.group()
+    @commands.guild_only()
+    async def set(self, ctx, *args):
+        """
+        Set a sound to play when you enter a voice channel.
+        Sound only plays if its been more than an hour since it last played. 
+        Example Usage: `entrance set fart long`
+        """
+        args = list(args)
+        if len(args) == 0:
+            group = 'root'
+            filename = args[0]
+            sound_id = filename
+        if len(args) == 1:
+            group = args[0]
+            filename = args[1]
+            sound_id = f"{group} {filename}"
+        try:
+            if not checkExists(group, filename):
+                raise Sound_Not_Exist_Error
+            member_id = ctx.message.author.id
+            db = GetDB(config.database_path)
+            db.cursor.execute("DELETE FROM entrance WHERE user_id=?",(member_id,))
+            db.cursor.execute("INSERT INTO entrance(sound_id, user_id, last_seen) VALUES(?,?,?)", (sound_id, member_id, "NULL"))
+            db.commit()
+            await ctx.message.author.send(format_markdown(f"Set entry sound to: \"{sound_id}\" for User {ctx.message.author.name}"))
+            db.close()
+            return
+
+        except Sound_Not_Exist_Error as e:
+            await ctx.message.author.send(format_markdown(e))
+            return
+        except Error as e:
+            await ctx.message.author.send(format_markdown("Something happened, please notify the bot owner."))
+            return
+
+    @entrance.group()
+    @commands.guild_only()
+    async def remove(self, ctx):
+        """
+        Remove the entry sound for the user. 
+        This will disable a sound playing when the user enters a voice chat. 
+        Example Usage: `entrance remove`
+        """
+        member_id = ctx.message.author.id
+        db = GetDB(config.database_path)
+        db.cursor.execute("DELETE FROM entrance WHERE user_id=?", (member_id,))
+        db.commit()
+        await ctx.message.author.send(format_markdown("Removed entry sound."))
+        db.close()
+        return
+
+    @entrance.group()
+    @commands.guild_only()
+    async def info(self, ctx):
+        """
+        DMs the user the entrance sound they have set.
+        Example Usage: `entrance info`
+        """
+        member_id = ctx.message.author.id
+        db = GetDB(config.database_path)
+        db.cursor.execute("SELECT sound_id FROM entrance WHERE user_id=?", (member_id,))
+        data = db.cursor.fetchall()
+        if len(data) == 0:
+            await ctx.message.author.send(format_markdown("You don't have an entry sound set."))
+            return
+        else:
+            await ctx.message.author.send(f"> {ctx.author.mention} your entrance sound is \"{data[0][0]}\"")
+        db.close()
 
 def setup(bot):
     bot.add_cog(Entrance(bot))
